+++ graphmaker/reports.py
@@ -5,9 +5,17 @@ from collections import Counter
 import networkx
 import numpy
 import scipy
-from graphmaker.data_utils import serializable_histogram
 
 # import planarity
+
+
+def serializable_histogram(data):
+    counts, bin_endpoints = numpy.histogram(data, bins='auto')
+    counts = list(map(int, counts))
+    bins = list()
+    for left, right in zip(bin_endpoints[:-1], bin_endpoints[1:]):
+        bins.append([round(float(left), 6), round(float(right), 6)])
+    return {'bins': bins, 'counts': counts}
 
 
 def edge_set(edges):
